플레이어 피격 구현

+++ Assets/ChildCollider.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 interface ITakeHit
 {
-    void OnTakeHit(int damage);
+    void OnTakeHit(float damage);
 }
 interface IReceiveMeleeAttackInfo
 {
